proper error reporting for internal error 30
This commit implements a proper exception which replaces "internal
error: 30", an error identifying the fact that a string of a rule has
reached 1 million matches. In other words, it's way too generic.
By reporting the rule name & string identifier users of software
using Yara should be able to fix their Yara rules (or at least be able
to quickly identify which rules are affected).