Remove python2 support.
Python2 is a dead language, we should no longer be supporting it. To that end,
this commit removes python2 support by doing the following:

If you run setup.py with a python2 interpreter, it will error out.

The pip runtime is now enforcing a required python version of 3 or higher.

Remove all preprocessor macros for supporting both python2 and python3. They are
all now python3 only.

I also removed the PYTHON_API_VERSION check since that is very old (as far as I
can tell that API version was first set in 1997).