Fix build on macos.
When building on macos I noticed that libcrypto was not found when using
has_function(). This is because the compiler used to compile the check program
is not being told about the library and include paths. Fix it by passing those
into the has_function() check where appropriate.

I have no idea how long this has been going on but it seems like it may have
been broken for a while now and we somehow didn't notice?