Release GIL while compiling Yara rules (#183)
* Release GIL when compiling rules.

The compilation is a blocking operation and can be potentially CPU
intensive. There is no reason to hold GIL.

* Lock the GIL in callback to prevent segmentation fault.

To be uniform with other callbacks and to prevent the SIGSEGV we have to
lock the GIL while we want to interact with python objects. The callback
is called from a C code releases GIL. Other callback functions use the same
locking strategy.