Always enable the dotnet module. (#222)
When building prior to this commit the dotnet module would be disabled unless
you explicitly asked for it. This is the opposite of what the default is in
libyara. Fix it by always building the dotnet module.

Tested with:

```
wxs@mbp yara-python % PYTHONPATH=build/lib.macosx-10.9-universal2-3.9 python3 -c 'import yara; print(yara.modules)'
['tests', 'pe', 'elf', 'math', 'time', 'console', 'string', 'dotnet', 'hash']
wxs@mbp yara-python %
```

Without this change the dotnet module would not be in the list.